fix(e12j): align GraphEdgeRow.confidence with SQL REAL (f32)
PostgreSQL REAL is FLOAT4 (4-byte), not FLOAT8 (8-byte). Changed:
- GraphEdgeRow.confidence: f64 → f32
- into_graph_edge(): self.confidence → self.confidence as f64

Also adds direct DecisionGraphExecutor tests:
- empty graph behavior (focus node, zero edges)
- focus-only behavior (no outgoing rationale edges)
- unsupported target behavior (rejects Symbol targets)
- shape-equality (preserves decision identity in view)

+++ crates/cognicode-explorer/src/adapters/pg_graph_repository.rs
@@ -545,7 +545,10 @@ struct GraphEdgeRow {
     target_id: String,
     kind: String,
     provenance: String,
-    confidence: f64,
+    /// confidence is REAL (FLOAT4) in PostgreSQL → decoded as f32.
+    /// Conversion to f64 happens inside into_graph_edge() when
+    /// constructing the domain GraphEdge.
+    confidence: f32,
     metadata: serde_json::Value,
 }
 
@@ -571,7 +574,7 @@ impl GraphEdgeRow {
             NodeId(self.target_id),
             kind,
             provenance,
-            self.confidence,
+            self.confidence as f64,
         )
         .map_err(|e| GraphError::Storage(format!("pg_graph_repository graph edge invalid: {e}")))?;
         edge.metadata = metadata;
